Move all schemaformwidget function definitions into one cpp
