+++ apps/AssetBuilder/src/form/schemaformwidget.cpp
@@ -24,17 +24,29 @@
 
 #include "form/schemaformwidget.h"
 
+#include "documentation.h"
+#include "form/collapsiblesection.h"
+#include "form/colorwidget.h"
 #include "form/matrixwidget.h"
+#include "form/searchdropdown.h"
+#include "identifierregistry.h"
+#include "schema/assetschema.h"
 #include "utils.h"
 #include <QCheckBox>
 #include <QComboBox>
+#include <QFileDialog>
+#include <QFrame>
 #include <QGridLayout>
 #include <QHBoxLayout>
 #include <QLabel>
 #include <QLineEdit>
 #include <QMessageBox>
 #include <QPushButton>
+#include <QRegularExpression>
+#include <QVBoxLayout>
 #include <algorithm>
+#include <functional>
+#include <limits>
 #include <optional>
 #include <string>
 #include <vector>
@@ -44,6 +56,48 @@ namespace {
     constexpr int ColumnSpacing = 8;
     constexpr int RowSpacing = 4;
     constexpr int InfoButtonSize = 16;
+    constexpr int CardMarginHorizontal = 8;
+    constexpr int CardMarginVertical = 6;
+    constexpr const char* DateDisplayFormat = "yyyy-MM-dd HH:mm:ss";
+
+    // Object names used as type tags for widget identification.
+    constexpr const char* FileContainerName = "file-container";
+    constexpr const char* UnionContainerName = "union-container";
+    constexpr const char* UnionTypeComboName = "union-type-combo";
+    constexpr const char* DateEditName = "date-edit";
+
+    enum class TableKind {
+        RefArray,     // e.g. Layers = [{ Type: "...", ... }, ...]
+        Ref,          // e.g. Renderable = { Type: "...", ... }
+        ScalarArray,  // e.g. Tags = ["tag1", "tag2"]
+        InlineArray,  // e.g. Instruments = [{ Name: "...", ... }, ...]
+        Table         // e.g. GUI = { Name: "...", Path: "..." }
+    };
+
+    struct TableClassification {
+        TableKind kind;
+        const SchemaReference* reference = nullptr;
+    };
+
+    constexpr MatrixTypeEntry MatrixTypeLookup[] = {
+    { "Vector2<double>",    2,  2, false, false },
+    { "Vector2<int>",       2,  2, true,  false },
+    { "Vector3<double>",    3,  3, false, false },
+    { "Vector3<int>",       3,  3, true,  false },
+    { "Vector4<double>",    4,  4, false, false },
+    { "Vector4<int>",       4,  4, true,  false },
+    { "Color3",             3,  3, false, true  },
+    { "Color4",             4,  4, false, true  },
+    { "Matrix3x3<double>",  9,  3, false, false },
+    { "Matrix4x4<double>", 16,  4, false, false }
+    };
+
+    struct NumericRange {
+        double min = -std::numeric_limits<double>::max();
+        double max = std::numeric_limits<double>::max();
+        bool hasMin = false;
+        bool hasMax = false;
+    };
 
     bool hasWidgetContent(QWidget* widget) {
         if (auto* checkBox = qobject_cast<QCheckBox*>(widget);  checkBox) {
@@ -286,6 +340,299 @@ namespace {
                 populateWidget(fieldWidget, value);
                 break;
             }
+        }
+    }
+
+    // Extracts numeric bounds from a schema description string. Supports "In range",
+    // "Greater or equal to", "Greater than", "Less or equal to", and "Less than" patterns
+    NumericRange parseRange(const std::string& description) {
+        NumericRange range;
+        const QString descriptionText = QString::fromStdString(description);
+
+        // "In range: ( min, max )"
+        static const QRegularExpression reRange(
+            R"(In range:\s*\(\s*([^,]+),\s*([^)]+)\))"
+        );
+        QRegularExpressionMatch match = reRange.match(descriptionText);
+        if (match.hasMatch()) {
+            bool okMin = false;
+            bool okMax = false;
+            const double lower = match.captured(1).trimmed().toDouble(&okMin);
+            const double upper = match.captured(2).trimmed().toDouble(&okMax);
+            if (okMin) {
+                range.min = lower;
+                range.hasMin = true;
+            }
+            if (okMax) {
+                range.max = upper;
+                range.hasMax = true;
+            }
+            return range;
+        }
+
+        // "Greater or equal to: N"
+        static const QRegularExpression reGe(R"(Greater or equal to:\s*([-\d.eE+]+))");
+        match = reGe.match(descriptionText);
+        if (match.hasMatch()) {
+            bool ok = false;
+            const double value = match.captured(1).toDouble(&ok);
+            if (ok) {
+                range.min = value;
+                range.hasMin = true;
+            }
+        }
+
+        // "Greater than: N"
+        static const QRegularExpression reGt(R"(Greater than:\s*([-\d.eE+]+))");
+        match = reGt.match(descriptionText);
+        if (match.hasMatch()) {
+            bool ok = false;
+            const double value = match.captured(1).toDouble(&ok);
+            if (ok) {
+                range.min = value;
+                range.hasMin = true;
+            }
+        }
+
+        // "Less or equal to: N"
+        static const QRegularExpression reLe(R"(Less or equal to:\s*([-\d.eE+]+))");
+        match = reLe.match(descriptionText);
+        if (match.hasMatch()) {
+            bool ok = false;
+            const double value = match.captured(1).toDouble(&ok);
+            if (ok) {
+                range.max = value;
+                range.hasMax = true;
+            }
+        }
+
+        // "Less than: N"
+        static const QRegularExpression reLt(R"(Less than:\s*([-\d.eE+]+))");
+        match = reLt.match(descriptionText);
+        if (match.hasMatch()) {
+            bool ok = false;
+            const double value = match.captured(1).toDouble(&ok);
+            if (ok) {
+                range.max = value;
+                range.hasMax = true;
+            }
+        }
+
+        return range;
+    }
+
+    // Builds a placeholder string describing the valid range for a numeric field. Returns
+    // e.g. "1 to 10", ">= 0", "<= 100", or a default "0"/"0.0"
+    QString rangePlaceholder(const NumericRange& range, bool isInteger) {
+        auto format = [isInteger](double value) {
+            if (isInteger) {
+                return QString::number(static_cast<int>(value));
+            }
+            QString formatted = QString::number(value, 'g', 10);
+            if (!formatted.contains('.')) {
+                formatted += ".0";
+            }
+            return formatted;
+            };
+
+        if (range.hasMin && range.hasMax) {
+            return format(range.min) + " to " + format(range.max);
+        }
+        if (range.hasMin) {
+            return ">= " + format(range.min);
+        }
+        if (range.hasMax) {
+            return "<= " + format(range.max);
+        }
+        return isInteger ? "0" : "0.0";
+    }
+
+    // Extracts the allowed values from an "In list { ... }" description pattern. Returns
+    // the values as a trimmed string list, or empty if no match
+    QStringList parseInList(const std::string& description) {
+        const QString descriptionText = QString::fromStdString(description);
+        static const QRegularExpression re(R"(In list \{\s*(.+?)\s*\})");
+        QRegularExpressionMatch match = re.match(descriptionText);
+        if (!match.hasMatch()) {
+            return QStringList();
+        }
+        QStringList items = match.captured(1).split(",");
+        for (QString& item : items) {
+            item = item.trimmed();
+        }
+        items.removeAll("");
+        return items;
+    }
+
+    // Splits a union type string like "Integer, or Double" into individual types. Handles
+    // both ", or " and " or " separators
+    QStringList splitUnionTypes(const QString& type) {
+        QString cleaned = type;
+        cleaned.replace(", or ", ", ");
+        cleaned.replace(" or ", ", ");
+        QStringList parts = cleaned.split(", ", Qt::SkipEmptyParts);
+        for (QString& part : parts) {
+            part = part.trimmed();
+        }
+        return parts;
+    }
+
+    // Determines the table kind (ref array, ref, scalar array, inline array, or plain
+// table) based on the member's structure and references
+    TableClassification classifyTableMember(const SchemaMember& member) {
+        // An array member has exactly one child named "*", which serves as the schema
+        // template for each element in the array
+        const bool isArray = !member.reference.has_value() &&
+            member.members.size() == 1 && member.members[0].name == "*";
+        const std::optional<SchemaReference>& optionalReference =
+            isArray ? member.members[0].reference : member.reference;
+
+        if (optionalReference.has_value() && optionalReference->isFound) {
+            return {
+                isArray ? TableKind::RefArray : TableKind::Ref,
+                &optionalReference.value()
+            };
+        }
+        if (isArray) {
+            return {
+                member.members[0].members.empty()
+                    ? TableKind::ScalarArray : TableKind::InlineArray,
+                nullptr
+            };
+        }
+        return { TableKind::Table, nullptr };
+    }
+
+    // Removes and deletes all widgets and layout items from a layout
+    void clearLayout(QLayout* layout) {
+        while (layout->count() > 0) {
+            QLayoutItem* layoutItem = layout->takeAt(0);
+            if (layoutItem->widget()) {
+                delete layoutItem->widget();
+            }
+            delete layoutItem;
+        }
+    }
+
+    // Looks up a schema type by display name within a category
+    const SchemaType* findTypeByName(const SchemaCategory* category,
+        const QString& typeName)
+    {
+        if (!category) {
+            return nullptr;
+        }
+        for (const SchemaType& schemaType : category->types) {
+            if (QString::fromStdString(schemaType.name) == typeName) {
+                return &schemaType;
+            }
+        }
+        return nullptr;
+    }
+
+    // Extracts the "Type" value from a property map, or empty if absent
+    QString storedTypeFromMap(const PropertyMap& properties) {
+        const auto it = properties.find("Type");
+        if (it != properties.end() && it->second.isString()) {
+            return QString::fromStdString(it->second.toString());
+        }
+        return {};
+    }
+
+    struct ArrayScaffold {
+        QVBoxLayout* itemsLayout;
+        QPushButton* addButton;
+    };
+
+    // Creates the shared container layout with an items area and "+ Add" button for
+    // array sections
+    ArrayScaffold createArrayScaffold(CollapsibleSection* section) {
+        // Outer container for the whole array section
+        QWidget* container = new QWidget();
+        QBoxLayout* containerLayout = new QVBoxLayout(container);
+        containerLayout->setContentsMargins(0, 4, 0, 4);
+        containerLayout->setSpacing(4);
+
+        // Inner container where individual item cards are added
+        QWidget* itemsContainer = new QWidget();
+        QVBoxLayout* itemsLayout = new QVBoxLayout(itemsContainer);
+        itemsLayout->setContentsMargins(0, 0, 0, 0);
+        itemsLayout->setSpacing(4);
+
+        // Button to append a new item to the array
+        QPushButton* addButton = new QPushButton("+ Add");
+        addButton->setObjectName("array-add-button");
+
+        // Assemble and set as the section content
+        containerLayout->addWidget(itemsContainer);
+        containerLayout->addWidget(addButton);
+        section->setContentWidget(container);
+
+        return { itemsLayout, addButton };
+    }
+
+    struct CardFrame {
+        QFrame* frame;
+        QBoxLayout* layout;
+    };
+
+    // Creates a styled card frame with vertical layout for an array item
+    CardFrame createCardFrame() {
+        QFrame* card = new QFrame();
+        card->setObjectName("array-item-frame");
+        card->setFrameShape(QFrame::StyledPanel);
+
+        QBoxLayout* cardLayout = new QVBoxLayout(card);
+
+        cardLayout->setContentsMargins(
+            CardMarginHorizontal,
+            CardMarginVertical,
+            CardMarginHorizontal,
+            CardMarginVertical
+        );
+        cardLayout->setSpacing(4);
+        return { card, cardLayout };
+    }
+
+    // The ensure* functions guarantee that a PropertyMap or PropertyList exists at the
+    // given key before a nested form or array binds to it. Without them, accessing an
+    // unset key would produce a null PropertyValue, and calling toMap()/toList() on it
+    // would fail
+
+    // Ensures the item at the given index is a map, replacing it if needed
+    PropertyMap& ensureListItemMap(PropertyList& items, size_t index) {
+        Q_ASSERT(index < items.size());
+        if (!items[index].isMap()) {
+            items[index] = PropertyValue{ PropertyMap{} };
+        }
+        return items[index].toMap();
+    }
+
+    // Ensures a property map exists at the given key, creating one if absent
+    PropertyMap& ensurePropertyMap(PropertyMap& properties, const std::string& key) {
+        if (properties.count(key) == 0 || properties.at(key).isNull()) {
+            properties[key] = PropertyValue{ PropertyMap{} };
+        }
+        Q_ASSERT(properties[key].isMap());
+        return properties[key].toMap();
+    }
+
+    // Ensures a property list exists at the given key. Creates one if absent, or wraps a
+    // single map value in a one-element list for uniformity
+    void ensurePropertyList(PropertyMap& properties, const std::string& key) {
+        if (properties.count(key) == 0 || properties.at(key).isNull()) {
+            properties[key] = PropertyValue{ PropertyList{} };
+        }
+        else if (properties.at(key).isMap()) {
+            // A single object was stored instead of a list - wrap it in a one-element
+            // list so downstream code can treat it uniformly as an array
+            PropertyMap existing = std::move(properties[key].toMap());
+            PropertyList wrapped;
+            wrapped.push_back(PropertyValue{ std::move(existing) });
+            properties[key] = PropertyValue{ std::move(wrapped) };
+        }
+        else {
+            // Should already be a list if it exists and isn't null or a map
+            Q_ASSERT(properties.at(key).isList());
         }
     }
 } // namespace
@@ -654,4 +1001,956 @@ void SchemaFormWidget::buildUi() {
     }
 
     outer->addWidget(gridWidget);
+}
+
+QWidget* SchemaFormWidget::createTableSection(const SchemaMember& member) {
+    // Create the collapsible section header
+    CollapsibleSection* section =
+        new CollapsibleSection(splitPascalCase(member.name), this);
+
+    section->setExpanded(_areSubSectionsExpanded);
+    section->setMandatory(!member.isOptional);
+
+    if (!member.documentation.empty()) {
+        section->setDocumentation(
+            QString::fromStdString(member.name),
+            QString::fromStdString(member.documentation)
+        );
+    }
+
+    // Property key for right-click copy/paste context menu
+    section->setSectionKey(QString::fromStdString(member.name));
+
+    // Forward section signals to the parent form
+    connect(
+        section, &CollapsibleSection::documentationRequested,
+        this, &SchemaFormWidget::documentationRequested
+    );
+    connect(
+        section, &CollapsibleSection::copyRequested,
+        this, &SchemaFormWidget::sectionCopyRequested
+    );
+    connect(
+        section, &CollapsibleSection::pasteRequested,
+        this, &SchemaFormWidget::sectionPasteRequested
+    );
+
+    // Populate the section content based on the table kind
+    const TableClassification classification = classifyTableMember(member);
+    switch (classification.kind) {
+    case TableKind::RefArray:
+        buildRefArrayContent(
+            section,
+            member,
+            classification.reference->identifier,
+            _properties
+        );
+        break;
+    case TableKind::Ref:
+        buildRefContent(section, member, *classification.reference, _properties);
+        break;
+    case TableKind::ScalarArray: {
+        // Scalar arrays are rendered as a single flat widget (e.g. a text field)
+        // rather than a list with add/remove buttons. member.members[0] is the "*"
+        // array template child — it has the right type but a placeholder name. Copy
+        // it and replace the name with the parent's so the flat widget displays and
+        // stores the value under the correct key (e.g. "Tags")
+        SchemaMember scalar = member.members[0];
+        scalar.name = member.name;
+        section->setContentWidget(createFlatWidget(scalar));
+        break;
+    }
+    case TableKind::InlineArray:
+        buildInlineArrayContent(section, member, _properties);
+        break;
+    case TableKind::Table:
+        buildInlineTableContent(section, member, _properties);
+        break;
+    }
+
+    return section;
+}
+
+// ---------------------------------------------------------------------------
+// Internal helpers
+// ---------------------------------------------------------------------------
+
+SchemaFormWidget* SchemaFormWidget::createNestedForm(
+    const std::vector<SchemaMember>& members,
+    PropertyMap& properties)
+{
+    SchemaFormWidget* inner = new SchemaFormWidget(
+        members,
+        properties,
+        nullptr,
+        false,
+        true,
+        _registry
+    );
+    connect(
+        inner, &SchemaFormWidget::fieldChanged,
+        this, &SchemaFormWidget::fieldChanged
+    );
+    connect(
+        inner, &SchemaFormWidget::documentationRequested,
+        this, &SchemaFormWidget::documentationRequested
+    );
+    connect(
+        inner, &SchemaFormWidget::browseJassetRequested,
+        this, &SchemaFormWidget::browseJassetRequested
+    );
+    return inner;
+}
+
+void SchemaFormWidget::emitTypeDocumentation(const SchemaCategory* category,
+    const QString& typeName)
+{
+    const SchemaType* schemaType = findTypeByName(category, typeName);
+    if (!schemaType) {
+        return;
+    }
+    const Documentation info = {
+        .name = splitPascalCase(schemaType->name),
+        .type = "Table",
+        .isOptional = false,
+        .documentation = QString::fromStdString(schemaType->description)
+    };
+    emit documentationRequested(info);
+}
+
+void SchemaFormWidget::rebuildItemForm(QBoxLayout* layout, PropertyMap& properties,
+    const QString& typeName,
+    const SchemaCategory* category)
+{
+    clearLayout(layout);
+    const SchemaType* type = findTypeByName(category, typeName);
+    if (!type) {
+        return;
+    }
+
+    properties["Type"] = PropertyValue{ typeName.toStdString() };
+    SchemaFormWidget* inner = createNestedForm(type->members, properties);
+    layout->addWidget(inner);
+    inner->populateFromProperties();
+}
+
+// ---------------------------------------------------------------------------
+// Array card builders
+// build*Card and rebuild* call each other: removing an item triggers a
+// rebuild, which clears all cards and re-creates them with updated indices.
+// ---------------------------------------------------------------------------
+void SchemaFormWidget::buildRefArrayCard(QBoxLayout* itemsLayout,
+    const std::string& memberName,
+    const SchemaCategory* category,
+    const std::string& referenceId, size_t index,
+    PropertyMap& properties)
+{
+    // Pointer alias - lambdas cannot capture references
+    PropertyMap* propertiesPtr = &properties;
+    PropertyList& items = properties[memberName].toList();
+    PropertyMap& itemProperties = ensureListItemMap(items, index);
+
+    // Card frame and header row
+    auto [card, cardLayout] = createCardFrame();
+
+    QWidget* headerRow = new QWidget();
+    QBoxLayout* headerLayout = new QHBoxLayout(headerRow);
+    headerLayout->setContentsMargins(0, 0, 0, 0);
+    headerLayout->setSpacing(4);
+
+    // Type dropdown - populated with concrete types, skipping the base type
+    SearchDropdown* dropdown = new SearchDropdown(card);
+    dropdown->setPlaceholderText("Select type...");
+
+    const QString storedType = storedTypeFromMap(itemProperties);
+
+    int preSelectedIndex = -1;
+    int dropdownIndex = 0;
+    if (category) {
+        for (const SchemaType& schemaType : category->types) {
+            // Skip the abstract base type of the category
+            if (schemaType.identifier == referenceId) {
+                continue;
+            }
+            const QString typeName = QString::fromStdString(schemaType.name);
+            dropdown->addItem(splitPascalCase(schemaType.name), typeName);
+            if (!storedType.isEmpty() && typeName == storedType) {
+                preSelectedIndex = dropdownIndex;
+            }
+            ++dropdownIndex;
+        }
+    }
+    dropdown->setCurrentIndex(preSelectedIndex);
+
+    // Remove button for this array item
+    QPushButton* removeButton = new QPushButton(RemoveGlyph, card);
+    removeButton->setObjectName("combo-clear-button");
+    removeButton->setFixedSize(RemoveButtonSize, RemoveButtonSize);
+    removeButton->setToolTip("Remove item");
+
+    // Assemble header: dropdown + remove button
+    headerLayout->addWidget(dropdown, 1);
+    headerLayout->addWidget(removeButton);
+    cardLayout->addWidget(headerRow);
+
+    // Inner form holder - populated when a type is selected
+    QWidget* innerHolder = new QWidget(card);
+    QBoxLayout* innerLayout = new QVBoxLayout(innerHolder);
+    innerLayout->setContentsMargins(0, 0, 0, 0);
+    innerLayout->setSpacing(0);
+    cardLayout->addWidget(innerHolder);
+
+    // Restore the inner form for a previously saved type selection
+    if (preSelectedIndex >= 0) {
+        rebuildItemForm(innerLayout, itemProperties, storedType, category);
+    }
+
+    // Type selected: reset the item's property map and rebuild the form
+    connect(
+        dropdown,
+        &SearchDropdown::activated,
+        this,
+        [this, propertiesPtr, memberName, index, innerLayout, category, dropdown]() {
+            PropertyMap& properties = *propertiesPtr;
+            PropertyList& list = properties[memberName].toList();
+            if (index >= list.size()) {
+                return;
+            }
+            list[index] = PropertyValue{ PropertyMap{} };
+            PropertyMap& itemProperties = list[index].toMap();
+            const QString typeName = dropdown->currentData().toString();
+            rebuildItemForm(innerLayout, itemProperties, typeName, category);
+            emit fieldChanged();
+        }
+    );
+
+    // Hover: show documentation for the highlighted type
+    connect(
+        dropdown,
+        &SearchDropdown::highlighted,
+        this,
+        [this, category](const QVariant& itemData) {
+            emitTypeDocumentation(category, itemData.toString());
+        }
+    );
+
+    // Remove button: erase this item and rebuild the entire array
+    connect(
+        removeButton,
+        &QPushButton::clicked,
+        this,
+        [this, propertiesPtr, memberName, index, itemsLayout, category, referenceId]() {
+            PropertyMap& properties = *propertiesPtr;
+            PropertyList& list = properties[memberName].toList();
+            if (index < list.size()) {
+                list.erase(list.begin() + static_cast<int>(index));
+            }
+            rebuildRefArray(itemsLayout, memberName, category, referenceId, properties);
+            emit fieldChanged();
+        }
+    );
+
+    itemsLayout->addWidget(card);
+}
+
+void SchemaFormWidget::rebuildRefArray(QBoxLayout* itemsLayout,
+    const std::string& memberName,
+    const SchemaCategory* category,
+    const std::string& referenceId,
+    PropertyMap& properties)
+{
+    clearLayout(itemsLayout);
+    PropertyList& items = properties[memberName].toList();
+    for (size_t i = 0; i < items.size(); i++) {
+        buildRefArrayCard(itemsLayout, memberName, category, referenceId, i, properties);
+    }
+}
+
+void SchemaFormWidget::buildInlineArrayCard(QBoxLayout* itemsLayout,
+    const std::string& memberName,
+    const std::vector<SchemaMember>& itemMembers,
+    size_t index, PropertyMap& properties)
+{
+    PropertyMap* propertiesPtr = &properties;
+    PropertyList& items = properties[memberName].toList();
+    PropertyMap& itemProperties = ensureListItemMap(items, index);
+
+    auto [card, cardLayout] = createCardFrame();
+
+    QWidget* headerRow = new QWidget();
+    QBoxLayout* headerLayout = new QHBoxLayout(headerRow);
+    headerLayout->setContentsMargins(0, 0, 0, 0);
+    headerLayout->setSpacing(4);
+    headerLayout->addStretch();
+
+    QPushButton* removeButton = new QPushButton(RemoveGlyph, card);
+    removeButton->setObjectName("combo-clear-button");
+    removeButton->setFixedSize(RemoveButtonSize, RemoveButtonSize);
+    removeButton->setToolTip("Remove item");
+    headerLayout->addWidget(removeButton);
+    cardLayout->addWidget(headerRow);
+
+    SchemaFormWidget* inner = createNestedForm(itemMembers, itemProperties);
+    cardLayout->addWidget(inner);
+    inner->populateFromProperties();
+
+    // Capture pointer - underlying vector lives in _members
+    const std::vector<SchemaMember>* membersPtr = &itemMembers;
+    // Remove button: erase this item and rebuild the array
+    connect(
+        removeButton,
+        &QPushButton::clicked,
+        this,
+        [this, propertiesPtr, memberName, index, itemsLayout, membersPtr]() {
+            PropertyMap& properties = *propertiesPtr;
+            PropertyList& list = properties[memberName].toList();
+            if (index < list.size()) {
+                list.erase(list.begin() + static_cast<int>(index));
+            }
+            rebuildInlineArray(itemsLayout, memberName, *membersPtr, properties);
+            emit fieldChanged();
+        }
+    );
+
+    itemsLayout->addWidget(card);
+}
+
+void SchemaFormWidget::rebuildInlineArray(QBoxLayout* itemsLayout,
+    const std::string& memberName,
+    const std::vector<SchemaMember>& itemMembers,
+    PropertyMap& properties)
+{
+    clearLayout(itemsLayout);
+    PropertyList& items = properties[memberName].toList();
+    for (size_t i = 0; i < items.size(); i++) {
+        buildInlineArrayCard(itemsLayout, memberName, itemMembers, i, properties);
+    }
+}
+
+// ---------------------------------------------------------------------------
+// Public content builders
+// ---------------------------------------------------------------------------
+void SchemaFormWidget::buildRefArrayContent(CollapsibleSection* section,
+    const SchemaMember& member,
+    const std::string& referenceId,
+    PropertyMap& properties)
+{
+    // Look up the category for the reference
+    const SchemaCategory* category =
+        AssetSchema::instance().findCategoryByTypeId(referenceId);
+    const std::string memberName = member.name;
+    PropertyMap* propertiesPtr = &properties;
+
+    auto [itemsLayout, addButton] = createArrayScaffold(section);
+
+    ensurePropertyList(properties, memberName);
+
+    // Build a card for each existing item in the array
+    PropertyList& items = properties[memberName].toList();
+    for (size_t i = 0; i < items.size(); i++) {
+        buildRefArrayCard(itemsLayout, memberName, category, referenceId, i, properties);
+    }
+
+    // "+ Add" button: appends a new item and builds its card
+    connect(
+        addButton,
+        &QPushButton::clicked,
+        this,
+        [this, propertiesPtr, memberName, itemsLayout, category, referenceId]() {
+            PropertyMap& properties = *propertiesPtr;
+            properties[memberName].toList().push_back(PropertyValue{ PropertyMap{} });
+            const size_t newIndex = properties[memberName].toList().size() - 1;
+            buildRefArrayCard(
+                itemsLayout,
+                memberName,
+                category,
+                referenceId,
+                newIndex,
+                properties
+            );
+            emit fieldChanged();
+        }
+    );
+}
+
+void SchemaFormWidget::buildRefContent(CollapsibleSection* section,
+    const SchemaMember& member,
+    const SchemaReference& reference,
+    PropertyMap& properties)
+{
+    // Look up the category and concrete type for the reference
+    const SchemaCategory* category =
+        AssetSchema::instance().findCategoryByTypeId(reference.identifier);
+    const SchemaType* targetType =
+        AssetSchema::instance().findType(reference.identifier);
+
+    // Both lookups should succeed if the schema is well-formed
+    Q_ASSERT(targetType);
+    Q_ASSERT(category);
+
+    // Polymorphic when the target is the base type of its category
+    // (e.g. "Renderable"), or as a fallback when lookup fails
+    const bool isPolymorphic = !targetType || !category ||
+        targetType->name == category->name;
+
+    if (isPolymorphic) {
+        buildPolymorphicRefContent(section, member, reference, category, properties);
+        return;
+    }
+
+    PropertyMap& subProperties = ensurePropertyMap(properties, member.name);
+    SchemaFormWidget* nested = createNestedForm(targetType->members, subProperties);
+
+    section->setContentWidget(nested);
+    nested->populateFromProperties();
+}
+
+void SchemaFormWidget::buildPolymorphicRefContent(CollapsibleSection* section,
+    const SchemaMember& member,
+    const SchemaReference& reference,
+    const SchemaCategory* category,
+    PropertyMap& properties)
+{
+    // Pointer alias - lambdas cannot capture references
+    PropertyMap* propertiesPtr = &properties;
+
+    // Outer container for the type selector and inner form
+    QWidget* container = new QWidget();
+    QBoxLayout* containerLayout = new QVBoxLayout(container);
+    containerLayout->setContentsMargins(0, 4, 0, 4);
+    containerLayout->setSpacing(4);
+
+    // Check if the property map already has a saved type selection
+    const bool hasStoredProperties =
+        properties.count(member.name) > 0
+        && properties.at(member.name).isMap();
+    const QString storedTypeName = hasStoredProperties ?
+        storedTypeFromMap(properties.at(member.name).toMap()) :
+        QString();
+
+    // Type dropdown - populated with all concrete types in the category, skipping the
+    // abstract base type and the member's own name
+    SearchDropdown* dropdown = new SearchDropdown(container);
+    dropdown->setPlaceholderText(
+        QString("No %1 selected").arg(splitPascalCase(member.name).toLower())
+    );
+    int preSelectedIndex = -1;
+    int dropdownIndex = 0;
+
+    // Add each concrete subtype to the dropdown
+    if (category) {
+        for (const SchemaType& schemaType : category->types) {
+            // Skip self-referential entry (e.g. "Renderable" member)
+            if (schemaType.name == member.name) {
+                continue;
+            }
+            // Skip the abstract base type of the category
+            if (schemaType.identifier == reference.identifier) {
+                continue;
+            }
+            const QString typeName = QString::fromStdString(schemaType.name);
+            dropdown->addItem(splitPascalCase(schemaType.name), typeName);
+
+            if (!storedTypeName.isEmpty() && typeName == storedTypeName) {
+                preSelectedIndex = dropdownIndex;
+            }
+            ++dropdownIndex;
+        }
+    }
+    dropdown->setCurrentIndex(preSelectedIndex);
+
+    // Inner form holder - populated when a type is selected
+    QWidget* innerHolder = new QWidget();
+    QBoxLayout* innerLayout = new QVBoxLayout(innerHolder);
+    innerLayout->setContentsMargins(0, 0, 0, 0);
+    innerLayout->setSpacing(0);
+
+    // Selector row: dropdown & clear button
+    QWidget* selectorRow = new QWidget();
+    QBoxLayout* selectorLayout = new QHBoxLayout(selectorRow);
+    selectorLayout->setContentsMargins(0, 0, 0, 0);
+    selectorLayout->setSpacing(4);
+    selectorLayout->addWidget(dropdown, 1);
+
+    QPushButton* clearButton = new QPushButton(RemoveGlyph);
+    clearButton->setObjectName("combo-clear-button");
+    clearButton->setFixedSize(RemoveButtonSize, RemoveButtonSize);
+    clearButton->setToolTip("Clear selection");
+    selectorLayout->addWidget(clearButton);
+    clearButton->setVisible(preSelectedIndex >= 0);
+
+    // Assemble the container and set it as the section content
+    containerLayout->addWidget(selectorRow);
+    containerLayout->addWidget(innerHolder);
+    section->setContentWidget(container);
+
+    const std::string memberName = member.name;
+
+    // Restore the inner form for a previously saved type selection
+    if (preSelectedIndex >= 0) {
+        PropertyMap& subProperties = ensurePropertyMap(properties, memberName);
+        rebuildItemForm(innerLayout, subProperties, storedTypeName, category);
+    }
+
+    // Type selected: reset the property map and rebuild the inner form
+    connect(
+        dropdown,
+        &SearchDropdown::activated,
+        this,
+        [this, propertiesPtr, memberName, innerLayout, dropdown, clearButton, category]()
+        {
+            PropertyMap& properties = *propertiesPtr;
+            properties[memberName] = PropertyValue{ PropertyMap{} };
+            PropertyMap& subProperties = properties[memberName].toMap();
+
+            rebuildItemForm(
+                innerLayout,
+                subProperties,
+                dropdown->currentData().toString(),
+                category
+            );
+            emit fieldChanged();
+            clearButton->setVisible(true);
+        }
+    );
+
+    // Hover: show documentation for the highlighted type
+    connect(
+        dropdown,
+        &SearchDropdown::highlighted,
+        this,
+        [this, category](const QVariant& itemData) {
+            emitTypeDocumentation(category, itemData.toString());
+        }
+    );
+
+    // Clear button: remove the selection and erase the property
+    connect(
+        clearButton,
+        &QPushButton::clicked,
+        this,
+        [this, propertiesPtr, memberName, dropdown, innerLayout, clearButton]() {
+            PropertyMap& properties = *propertiesPtr;
+            dropdown->setCurrentIndex(-1);
+            clearButton->setVisible(false);
+            clearLayout(innerLayout);
+            properties.erase(memberName);
+            emit fieldChanged();
+        }
+    );
+}
+
+void SchemaFormWidget::buildInlineArrayContent(CollapsibleSection* section,
+    const SchemaMember& member,
+    PropertyMap& properties)
+{
+    // members[0] is the "*" array template child; its sub-members define the schema for
+    // each element in the array
+    const std::vector<SchemaMember>& itemMembers = member.members[0].members;
+    const std::string memberName = member.name;
+    const std::vector<SchemaMember>* membersPtr = &itemMembers;
+    PropertyMap* propertiesPtr = &properties;
+
+    auto [itemsLayout, addButton] = createArrayScaffold(section);
+
+    ensurePropertyList(properties, memberName);
+
+    PropertyList& items = properties[memberName].toList();
+    for (size_t i = 0; i < items.size(); i++) {
+        buildInlineArrayCard(itemsLayout, memberName, *membersPtr, i, properties);
+    }
+
+    // "+ Add" button: appends a new item to the array and builds its card
+    connect(
+        addButton,
+        &QPushButton::clicked,
+        this,
+        [this, propertiesPtr, memberName, itemsLayout, membersPtr]() {
+            PropertyMap& properties = *propertiesPtr;
+            properties[memberName].toList().push_back(PropertyValue{ PropertyMap{} });
+            const size_t newIndex = properties[memberName].toList().size() - 1;
+            buildInlineArrayCard(
+                itemsLayout,
+                memberName,
+                *membersPtr,
+                newIndex,
+                properties
+            );
+            emit fieldChanged();
+        }
+    );
+}
+
+void SchemaFormWidget::buildInlineTableContent(CollapsibleSection* section,
+    const SchemaMember& member,
+    PropertyMap& properties)
+{
+    if (member.members.empty()) {
+        return;
+    }
+
+    PropertyMap& subProperties = ensurePropertyMap(properties, member.name);
+    SchemaFormWidget* nested = createNestedForm(member.members, subProperties);
+    section->setContentWidget(nested);
+    nested->populateFromProperties();
+}
+
+QWidget* SchemaFormWidget::createFlatWidget(const SchemaMember& member) {
+    const QString typeString = QString::fromStdString(member.type);
+
+    // Union types (contain ", " or " or ")
+    if (typeString.contains(", ") || typeString.contains(" or ")) {
+        return createUnionWidget(member);
+    }
+
+    if (member.type == "Boolean") {
+        return createBooleanWidget(member.name);
+    }
+    if (member.type == "Integer") {
+        return createIntegerWidget(member.name, member.description);
+    }
+    if (member.type == "Double") {
+        return createDoubleWidget(member.name, member.description);
+    }
+    for (const MatrixTypeEntry& entry : MatrixTypeLookup) {
+        if (member.type == entry.typeName) {
+            return createMatrixWidget(member.name, entry);
+        }
+    }
+    if (member.type == "File") {
+        return createFileWidget(member.name, member.description, false);
+    }
+    if (member.type == "Directory") {
+        return createFileWidget(member.name, member.description, true);
+    }
+    if (member.type == "Date and time") {
+        return createDateTimeWidget(member.name);
+    }
+    // Identifier fields that reference other nodes (not the node's own Identifier
+    // definition) get a registry combo + browse button
+    if (member.type == "Identifier" && member.name != "Identifier" && _registry) {
+        return createIdentifierComboWidget(member);
+    }
+    const QStringList options = parseInList(member.description);
+    if (!options.isEmpty()) {
+        return createInListWidget(member.name, options);
+    }
+
+    // Fallback to string
+    return createStringWidget(member);
+}
+
+QWidget* SchemaFormWidget::createBooleanWidget(const std::string& name) {
+    QCheckBox* checkBox = new QCheckBox(this);
+
+    connect(
+        checkBox,
+        &QCheckBox::toggled,
+        this,
+        [this, name](bool checked) {
+            _properties[name] = PropertyValue{ checked };
+            emit fieldChanged();
+        }
+    );
+    return checkBox;
+}
+
+QWidget* SchemaFormWidget::createIntegerWidget(const std::string& name,
+    const std::string& description)
+{
+    QLineEdit* lineEdit = new QLineEdit(this);
+    const NumericRange range = parseRange(description);
+    // Cannot cast the NumericRange double defaults to int (overflow / UB), so fall back
+    // to int limits when no bound was parsed
+    const int lowerBound =
+        range.hasMin ? static_cast<int>(range.min) : std::numeric_limits<int>::min();
+    const int upperBound =
+        range.hasMax ? static_cast<int>(range.max) : std::numeric_limits<int>::max();
+    lineEdit->setValidator(new QIntValidator(lowerBound, upperBound, lineEdit));
+    lineEdit->setPlaceholderText(rangePlaceholder(range, true));
+
+    connect(
+        lineEdit,
+        &QLineEdit::textEdited,
+        this,
+        [this, name](const QString& text) {
+            bool ok = false;
+            const int value = text.toInt(&ok);
+            _properties[name] = PropertyValue{
+                static_cast<double>(ok ? value : 0)
+            };
+            emit fieldChanged();
+        }
+    );
+    return lineEdit;
+}
+
+QWidget* SchemaFormWidget::createDoubleWidget(const std::string& name,
+    const std::string& description)
+{
+    QLineEdit* lineEdit = new QLineEdit(this);
+    const NumericRange range = parseRange(description);
+    // QDoubleValidator misbehaves with extreme double values, so use +/- 1e15 as a
+    // practical fallback when no bound was parsed
+    const double lowerBound = range.hasMin ? range.min : -1e15;
+    const double upperBound = range.hasMax ? range.max : 1e15;
+    lineEdit->setValidator(new QDoubleValidator(lowerBound, upperBound, -1, lineEdit));
+    lineEdit->setPlaceholderText(rangePlaceholder(range, false));
+    connect(
+        lineEdit,
+        &QLineEdit::textEdited,
+        this,
+        [this, name](const QString& text) {
+            bool ok = false;
+            const double value = text.toDouble(&ok);
+            _properties[name] = PropertyValue{ ok ? value : 0.0 };
+            emit fieldChanged();
+        }
+    );
+    return lineEdit;
+}
+
+QWidget* SchemaFormWidget::createMatrixWidget(const std::string& name,
+    const MatrixTypeEntry& entry)
+{
+    MatrixWidget* matrixWidget = entry.isColor ?
+        static_cast<MatrixWidget*>(new ColorWidget(entry.nComponents, this)) :
+        new MatrixWidget(entry.nComponents, entry.nColumns, entry.isInteger, this);
+    connect(
+        matrixWidget,
+        &MatrixWidget::valueChanged,
+        this,
+        [this, name, matrixWidget]() {
+            _properties[name] = PropertyValue{ matrixWidget->values() };
+            emit fieldChanged();
+        }
+    );
+    return matrixWidget;
+}
+
+QWidget* SchemaFormWidget::createFileWidget(const std::string& name,
+    const std::string& description,
+    bool isDirectory)
+{
+    QWidget* container = new QWidget(this);
+    container->setObjectName(FileContainerName);
+    QBoxLayout* horizontalLayout = new QHBoxLayout(container);
+    horizontalLayout->setContentsMargins(0, 0, 0, 0);
+    horizontalLayout->setSpacing(4);
+
+    QLineEdit* lineEdit = new QLineEdit(container);
+    if (!description.empty()) {
+        lineEdit->setPlaceholderText(QString::fromStdString(description));
+    }
+    else {
+        lineEdit->setPlaceholderText(
+            isDirectory ? "Path to directory..." : "Path to file..."
+        );
+    }
+    connect(
+        lineEdit,
+        &QLineEdit::textEdited,
+        this,
+        [this, name](const QString& text) {
+            _properties[name] = PropertyValue{ text.toStdString() };
+            emit fieldChanged();
+        }
+    );
+
+    QPushButton* browseButton = new QPushButton("Browse...", container);
+    browseButton->setObjectName("file-browse-button");
+    connect(
+        browseButton,
+        &QPushButton::clicked,
+        this,
+        [this, lineEdit, name, isDirectory]() {
+            const QString path = isDirectory ?
+                QFileDialog::getExistingDirectory(this, "Select Directory") :
+                QFileDialog::getOpenFileName(this, "Select File");
+            if (!path.isEmpty()) {
+                lineEdit->setText(path);
+                _properties[name] = PropertyValue{ path.toStdString() };
+                emit fieldChanged();
+            }
+        }
+    );
+
+    horizontalLayout->addWidget(lineEdit, 1);
+    horizontalLayout->addWidget(browseButton);
+    return container;
+}
+
+QWidget* SchemaFormWidget::createDateTimeWidget(const std::string& name) {
+    QLineEdit* lineEdit = new QLineEdit(this);
+    lineEdit->setObjectName(DateEditName);
+    // Human-readable hint; the actual Qt format string is DateDisplayFormat
+    lineEdit->setPlaceholderText("YYYY-MM-DD hh:mm:ss");
+    lineEdit->setValidator(new QRegularExpressionValidator(
+        QRegularExpression(R"(^\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2}$)"),
+        lineEdit
+    ));
+    connect(
+        lineEdit,
+        &QLineEdit::textEdited,
+        this,
+        [this, name](const QString& text) {
+            const QDateTime dateTime = QDateTime::fromString(text, DateDisplayFormat);
+            if (dateTime.isValid()) {
+                _properties[name] = PropertyValue{
+                    dateTime.toUTC().toString(Qt::ISODate).toStdString()
+                };
+            }
+            emit fieldChanged();
+        }
+    );
+    return lineEdit;
+}
+
+QWidget* SchemaFormWidget::createIdentifierComboWidget(const SchemaMember& member) {
+    QWidget* container = new QWidget(this);
+    QBoxLayout* horizontalLayout = new QHBoxLayout(container);
+    horizontalLayout->setContentsMargins(0, 0, 0, 0);
+    horizontalLayout->setSpacing(4);
+
+    // Returns identifiers excluding this node's own
+    auto filteredIdentifiers = [this]() {
+        QStringList identifiers = _registry->knownIdentifiers();
+        const auto it = _properties.find("Identifier");
+        if (it != _properties.end() && it->second.isString()) {
+            identifiers.removeAll(QString::fromStdString(it->second.toString()));
+        }
+        return identifiers;
+        };
+
+    QComboBox* combo = new QComboBox(container);
+    combo->setEditable(true);
+    combo->addItems(filteredIdentifiers());
+    if (!member.description.empty()) {
+        combo->lineEdit()->setPlaceholderText(QString::fromStdString(member.description));
+    }
+
+    connect(
+        _registry,
+        &IdentifierRegistry::registryChanged,
+        combo,
+        [combo, filteredIdentifiers]() {
+            const QString current = combo->currentText();
+            combo->blockSignals(true);
+            combo->clear();
+            combo->addItems(filteredIdentifiers());
+            combo->setCurrentText(current);
+            combo->blockSignals(false);
+        }
+    );
+    connect(
+        combo,
+        &QComboBox::currentTextChanged,
+        this,
+        [this, name = member.name](const QString& text) {
+            _properties[name] = PropertyValue{ text.toStdString() };
+            emit fieldChanged();
+        }
+    );
+
+    QPushButton* browseButton = new QPushButton("Browse .jasset", container);
+    browseButton->setObjectName("identifier-browse-button");
+    connect(
+        browseButton, &QPushButton::clicked,
+        this, &SchemaFormWidget::browseJassetRequested
+    );
+
+    horizontalLayout->addWidget(combo, 1);
+    horizontalLayout->addWidget(browseButton);
+    return container;
+}
+
+QWidget* SchemaFormWidget::createInListWidget(const std::string& name,
+    const QStringList& listOptions)
+{
+    QComboBox* combo = new QComboBox(this);
+    combo->addItem("Select...");
+    combo->addItems(listOptions);
+    combo->setCurrentIndex(0);
+    connect(
+        combo,
+        &QComboBox::currentIndexChanged,
+        this,
+        [this, name, combo](int index) {
+            // Index 0 is the "Select..." placeholder - erase the property so the
+            // placeholder choice is not serialized into the asset
+            if (index <= 0) {
+                _properties.erase(name);
+            }
+            else {
+                _properties[name] = PropertyValue{ combo->currentText().toStdString() };
+            }
+            emit fieldChanged();
+        }
+    );
+    return combo;
+}
+
+QWidget* SchemaFormWidget::createUnionWidget(const SchemaMember& member) {
+    const QString typeString = QString::fromStdString(member.type);
+    QStringList types = splitUnionTypes(typeString);
+
+    // Combo selector + show/hide pages. Hidden widgets take zero layout
+    // space, so the container fits the active page.
+    QWidget* container = new QWidget(this);
+    container->setObjectName(UnionContainerName);
+    QBoxLayout* verticalLayout = new QVBoxLayout(container);
+    verticalLayout->setContentsMargins(0, 0, 0, 0);
+    verticalLayout->setSpacing(4);
+
+    QComboBox* typeCombo = new QComboBox(container);
+    typeCombo->setObjectName(UnionTypeComboName);
+    verticalLayout->addWidget(typeCombo);
+
+    QList<QWidget*> typeWidgets;
+    for (int j = 0; j < types.size(); j++) {
+        typeCombo->addItem(types[j]);
+        // Empty description — the union description applies to the union
+        // as a whole, not the individual type
+        SchemaMember typeMember = { member.name, types[j].toStdString(), {} };
+        QWidget* widget = (types[j] == "Table") ?
+            createTableSection(typeMember) :
+            createFlatWidget(typeMember);
+
+        // Only the first type is visible; the combo switches between them
+        widget->setVisible(j == 0);
+        verticalLayout->addWidget(widget);
+        typeWidgets.append(widget);
+    }
+
+    connect(
+        typeCombo,
+        &QComboBox::currentIndexChanged,
+        this, [typeWidgets](int index) {
+            for (int i = 0; i < typeWidgets.size(); ++i) {
+                typeWidgets[i]->setVisible(i == index);
+            }
+        }
+    );
+
+    return container;
+}
+
+QWidget* SchemaFormWidget::createStringWidget(const SchemaMember& member) {
+    QLineEdit* lineEdit = new QLineEdit(this);
+
+    if (!member.description.empty()) {
+        lineEdit->setPlaceholderText(QString::fromStdString(member.description));
+    }
+
+    connect(
+        lineEdit,
+        &QLineEdit::textEdited,
+        this,
+        [this, name = member.name](const QString& text) {
+            _properties[name] = PropertyValue{ text.toStdString() };
+            emit fieldChanged();
+        }
+    );
+
+    return lineEdit;
 }
